moved components to files and added time to display component
